requirements.txt
pip==25.0
django-jazzmin==3.0.1
django-pgcrypto==3.0.3
whitenoise==6.8.2
Django==5.1.5
python-dotenv==1.0.1
psycopg2-binary==2.9.10
beautifulsoup4==4.13.2
pillow==11.1.0
